Create p12
Implement a function, called is_credit_card_valid(number), which returns True or False based on the following algorithm:

Starting from the right, double every second digit.
If after doubling the digit becomes >= 10, calculate the sum of the digits of the new number & use that. For example, if we have 9 and doubling it gets us 18, the final result is going to be 1 + 8 = 9
After the transformation, we find the sum of all digits in the transformed number.
The number is valid, if the sum is divisible by 10.
This is also known as the Luhn algorithm.