Create p16
Each key contains some letters from the alphabet. And by pressing that key, you rotate the letters until you get to your desired one.
It's time to implement some encode / decode functions for the old keypads!
First, implement a function that takes a list of integers - the sequence of numbers that have been pressed. The returned value should be the corresponding string of the message.
If you press 1, the next letter is going to be capitalized
If you press 0, this will insert a single white-space
If you press a number and wait for a few seconds, the special breaking number -1 enters the sequence. This is the way to write different letters from only one keypad!