Create p19
Ivan and his friends love going to the cinema (that's pre-pandemic).
But when it comes to picking seats, they don't like the usual & boring "next to each other in one row" placement.
That's why they always pick their seats in a strange way:
Maria should be above Ivan.
Georgi should be to the right of Maria.
Veronika should be above Maria.
And when they are done, they can take really strange forms (at least, strange in the eyes of the cinema owners).
You'll be given 2 arguments:
The cinema layout
The configuration our friends want to take.
The function should return a list of all possible placements, that satisfy the given form. This is basically a list of possible cinema layouts (structure the layouts the same way as the input)
Possible placement is a configuration where:
Our friends can book seats in the way they want.
They are not going outside of the cinema.
They are not taking any already reserved seats.